#4 employe
Creation entite, DTO etc...

+++ src/main/java/com/aclab/dne/repositories/PersonneRepository.java
@@ -3,5 +3,5 @@ package com.aclab.dne.repositories;
 import com.aclab.dne.model.Personne;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PersonneRepository extends CrudRepository<Personne,Integer> {
+public interface PersonneRepository extends CrudRepository<Personne,Long> {
 }
